+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.0.1.0
 
 - Binary operator options can now be parsed/formatted as JSON in a more compact
-  way: `foo --times bar --times baz` is can now be serialized/deserialized as
+  way: `foo --times bar --times baz` can now be serialized/deserialized as
   `{"times":[foo,bar,baz]}`
 - `--package PACKAGE` (i.e. without any version or version range specified) now
   stands for all versions of `PACKAGE`. Likewise in JSON